TimePicker: add DayChangeCallback interface.
This interface has function changeDay(direction: Int).
`direction` specifies the number of days to advance (+1 or -1).
The interface user should handle cases of 0 as no-ops.

When the hour rolls over in a way that signifies a change in date,
it will call this callback with the appropriate direction, if set.
This is helpful for combined date-time pickers.